Refactor dashboard into components
